

\name{ compute.llik.c }
\alias{ compute.llik.c }
\title{ compute.llik.c }

\description{
  TODO
}

\usage{
  compute.llik.c(X, y, weights, beta, intercept=TRUE)
}

\arguments{
  \item{ X }{
    TBD
  }


  \item{ y }{TBD
  }
  
  \item{ weights }{TBD  
  }
  \item{ beta }{
    TBD
  }
  \item{ intercept }{
    TBD
  }
}


\seealso{
  TBD
  %% \code{  \link{ help } }  
}

\details{
  TBD
}

\value{
  return value TBD
}

\references{
  TBD
}

%\author{
%  TBD
%}

\note{
  TBD
}

\examples{
  #TBD
}

\keyword{ TBD }

