

\name{ gen.esnps }
\alias{ gen.esnps }
\title{ gen.esnps }

\description{
  Curate the objective data to perform pathway analysis.
}

\usage{
  gen.esnps(snp.dat, snp.info, gene.info, gene.set, y, weights = NULL,
  gene.def=c("abs", "rel"), k=1, dist=20000)
}

\arguments{
  \item{ snp.dat }{
    TBD
  }

  \item{ snp.info }{ 
    TBD}

  \item{ gene.info }{ 
    TBD}

  \item{ gene.set }{ 
    TBD}

  \item{ y }{
    TBD
  }

  \item{ weights }{
    TBD
  }


  \item{ gene.def }{
    TBD
  }

  \item{ k }{
    TBD
  }

  \item{ dist }{
    TBD
  }

}

\seealso{
  TBD
  %% \code{  \link{ help } }  
}

\details{
TBD}

\value{
 TBD
}

\references{

  TBD
}

\note{
  TBD
}

\examples{
  #TBD
}

\keyword{ TBD }

