

\name{ Gene2Set }
\alias{ Gene2Set }
\title{ Gene2Set }

\description{
  Convert gene names in a lis of pathways to indices
}

\usage{
  Gene2Set(gene, set)
}

\arguments{
  \item{ gene }{
    TBD
  }

  \item{ set }{ TBD
  }
}

\seealso{
  TBD
  %% \code{  \link{ help } }  
}

\details{
  TBD
}

\value{
  TBD
}

\references{
  TBD
}

%\author{
%  TBD
%}

\note{
  TBD
}

\examples{
  #TBD
}

\keyword{ TBD }

