

\name{ logistic.ridge.c }
\alias{ logistic.ridge.c }
\title{ logistic.ridge.c }

\description{
  Find minima of penalized likelihood function
}

\usage{
logistic.ridge.c(X, y, grp, weights=NULL, lambda=1, tol=0.001, maxiter=100, 
intercept=TRUE, eq.eSNP=TRUE)
}

\arguments{
  \item{ X }{
   TBD
  }

  \item{ y }{
    TBD
  }

  \item{ grp }{TBD }

  
  \item{ weights }{TBD
  }

  \item{ lambda }{
    TBD
  }

  \item{ tol }{
    TBD
  }

  \item{ maxiter }{
    TBD
  }

  \item{ intercept }{ TBD}

  \item{ eq.eSNP }{ TBD }

}


\seealso{
  TBD
  %% \code{  \link{ help } }  
}

\details{
  TBD
}

\value{
  TBD
}

\references{
  TBD
}

%\author{
%  TBD
%}

\note{
  TBD
}

\examples{
  #TBD
}

\keyword{ TBD }

